refactor: enhance README and yolo26-pose-tracker.py for clarity and functionality; remove yolo26-pose.py
