+++ examples/YOLO-Axelera-Python/README.md
@@ -1,6 +1,6 @@
 # Ultralytics YOLO on Axelera Voyager SDK
 
-This document explains how to build and run complete inference pipelines on Axelera Metis devices using the Axelera runtime.
+This document explains how to build and run complete Ultralytics YOLO inference pipelines on Axelera Metis devices using the Axelera runtime.
 
 The typical workflow has two phases:
 
@@ -25,13 +25,12 @@ poses = pipeline(frame)  # frame in, results out
 
 ## Examples
 
-Three examples are provided below, ordered from simple to advanced. It should be straightforward to apply them to other models and tasks.
+Two examples are provided below. It should be straightforward to apply them to other models and tasks.
 
-| Script                   | Task                                    | Model                   | Description                                                           |
-| ------------------------ | --------------------------------------- | ----------------------- | --------------------------------------------------------------------- |
-| `yolo26-pose.py`         | Pose estimation — 17 COCO keypoints     | YOLO26n-pose (NMS-free) | Start here. Linear `op.seq()` pipeline with a custom operator.        |
-| `yolo11-seg.py`          | Instance segmentation                   | YOLO11n-seg             | Branching with `op.par()` for multi-head models (detections + masks). |
-| `yolo26-pose-tracker.py` | Pose estimation + multi-object tracking | YOLO26n-pose (NMS-free) | Adds stateful `op.tracker()` for multi-object tracking.               |
+| Script                   | Task                                                 | Model                   | Description                                                           |
+| ------------------------ | ---------------------------------------------------- | ----------------------- | --------------------------------------------------------------------- |
+| `yolo26-pose-tracker.py` | Pose estimation with optional multi-object tracking  | YOLO26n-pose (NMS-free) | Linear `op.seq()` pipeline with a custom operator and optional `op.tracker()`. Use `--tracker none` for pose-only mode. |
+| `yolo11-seg.py`          | Instance segmentation                                | YOLO11n-seg             | Branching with `op.par()` for multi-head models (detections + masks). |
 
 ## Quick Start
 
@@ -68,22 +67,22 @@ The compiled models are saved to `yolo26n-pose_axelera_model/` and
 
 ### Run
 
-#### Pose Estimation (YOLO26)
+#### Pose Estimation (Ultralytics YOLO26)
 
 ```bash
-python yolo26-pose.py --model yolo26n-pose.axm --source 0         # webcam
-python yolo26-pose.py --model yolo26n-pose.axm --source video.mp4 # video
-python yolo26-pose.py --model yolo26n-pose.axm --source image.jpg # image
+python yolo26-pose-tracker.py --model yolo26n-pose.axm --source 0 --tracker none  # pose only # webcam
+python yolo26-pose-tracker.py --model yolo26n-pose.axm --source image.jpg --tracker none # pose only # image
+python yolo26-pose-tracker.py --model yolo26n-pose.axm --source video.mp4 --tracker none # pose only # video
 ```
 
-#### Pose Tracking (YOLO26 + TrackTrack)
+#### Pose Tracking (Ultralytics YOLO26 + TrackTrack)
 
 ```bash
-python yolo26-pose-tracker.py --model yolo26n-pose.axm --source video.mp4
-python yolo26-pose-tracker.py --model yolo26n-pose.axm --source 0 --tracker bytetrack
+python yolo26-pose-tracker.py --model yolo26n-pose.axm --source video.mp4 # video with tracking
+python yolo26-pose-tracker.py --model yolo26n-pose.axm --source 0 --tracker tracktrack # webcam with tracking
 ```
 
-#### Instance Segmentation (YOLO11)
+#### Instance Segmentation (Ultralytics YOLO11)
 
 ```bash
 python yolo11-seg.py --model yolo11n-seg.axm --source 0
@@ -92,13 +91,13 @@ python yolo11-seg.py --model yolo11n-seg.axm --source video.mp4 --conf 0.3 --iou
 
 ### Arguments
 
-| Argument    | Default      | Description                                                                            |
-| ----------- | ------------ | -------------------------------------------------------------------------------------- |
-| `--model`   | _required_   | Path to compiled `.axm` model                                                          |
-| `--source`  | `0`          | Image path, video path, or webcam index                                                |
-| `--conf`    | `0.25`       | Confidence threshold                                                                   |
-| `--iou`     | `0.45`       | NMS IoU threshold _(segmentation only)_                                                |
-| `--tracker` | `tracktrack` | Tracking algorithm: `bytetrack`, `oc-sort`, `sort`, `tracktrack` _(pose-tracker only)_ |
+| Argument    | Default      | Description                                                              |
+| ----------- | ------------ | ------------------------------------------------------------------------ |
+| `--model`   | _required_   | Path to compiled `.axm` model                                            |
+| `--source`  | `0`          | Image path, video path, or webcam index                                  |
+| `--conf`    | `0.25`       | Confidence threshold                                                     |
+| `--iou`     | `0.45`       | NMS IoU threshold _(segmentation only)_                                  |
+| `--tracker` | `tracktrack` | Tracking algorithm: `bytetrack`, `oc-sort`, `sort`, `tracktrack`, `none` |
 
 ## What You'll See in the Code
